feat: Update LED index handling in reconnectStrips and ensure integer type for UI schema values

+++ src/JaamWeb.cpp
@@ -2434,7 +2434,7 @@ void JaamWeb::handleUiSchema() {
             if (i > 0) result += ",";
             result += fieldName;
             result += operand;
-            result += values[i];
+            result += static_cast<int>(values[i]);
         }
         return result;
     };
